show appointments on left
