+++ components/calendar/NewCalendar.tsx
@@ -1,254 +1,27 @@
-import {
-  CalendarIcon,
-  ChevronLeftIcon,
-  ChevronRightIcon,
-  EllipsisHorizontalIcon,
-  MapPinIcon,
-} from '../heroicons.tsx'
-// import { Menu, Transition } from '@headlessui/react'
-import { Maybe } from '../../types.ts'
+import Calendar from './Calendar.tsx'
+import { HealthWorkerAppointment } from '../../types.ts'
+import Appointment from './Appointment.tsx'
 
-const meetings = [
-  {
-    id: 1,
-    date: 'January 10th, 2022',
-    time: '5:00 PM',
-    datetime: '2022-01-10T17:00',
-    name: 'Leslie Alexander',
-    imageUrl:
-      'https://images.unsplash.com/photo-1494790108377-be9c29b29330?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80',
-    location: 'Starbucks',
-  },
-  // More meetings...
-]
-const days = [
-  { date: '2021-12-27' },
-  { date: '2021-12-28' },
-  { date: '2021-12-29' },
-  { date: '2021-12-30' },
-  { date: '2021-12-31' },
-  { date: '2022-01-01', isCurrentMonth: true },
-  { date: '2022-01-02', isCurrentMonth: true },
-  { date: '2022-01-03', isCurrentMonth: true },
-  { date: '2022-01-04', isCurrentMonth: true },
-  { date: '2022-01-05', isCurrentMonth: true },
-  { date: '2022-01-06', isCurrentMonth: true },
-  { date: '2022-01-07', isCurrentMonth: true },
-  { date: '2022-01-08', isCurrentMonth: true },
-  { date: '2022-01-09', isCurrentMonth: true },
-  { date: '2022-01-10', isCurrentMonth: true },
-  { date: '2022-01-11', isCurrentMonth: true },
-  { date: '2022-01-12', isCurrentMonth: true, isToday: true },
-  { date: '2022-01-13', isCurrentMonth: true },
-  { date: '2022-01-14', isCurrentMonth: true },
-  { date: '2022-01-15', isCurrentMonth: true },
-  { date: '2022-01-16', isCurrentMonth: true },
-  { date: '2022-01-17', isCurrentMonth: true },
-  { date: '2022-01-18', isCurrentMonth: true },
-  { date: '2022-01-19', isCurrentMonth: true },
-  { date: '2022-01-20', isCurrentMonth: true },
-  { date: '2022-01-21', isCurrentMonth: true },
-  { date: '2022-01-22', isCurrentMonth: true, isSelected: true },
-  { date: '2022-01-23', isCurrentMonth: true },
-  { date: '2022-01-24', isCurrentMonth: true },
-  { date: '2022-01-25', isCurrentMonth: true },
-  { date: '2022-01-26', isCurrentMonth: true },
-  { date: '2022-01-27', isCurrentMonth: true },
-  { date: '2022-01-28', isCurrentMonth: true },
-  { date: '2022-01-29', isCurrentMonth: true },
-  { date: '2022-01-30', isCurrentMonth: true },
-  { date: '2022-01-31', isCurrentMonth: true },
-  { date: '2022-02-01' },
-  { date: '2022-02-02' },
-  { date: '2022-02-03' },
-  { date: '2022-02-04' },
-  { date: '2022-02-05' },
-  { date: '2022-02-06' },
-]
-
-function classNames(...classes: Maybe<string>[]) {
-  return classes.filter(Boolean).join(' ')
-}
-
-export default function NewCalendar() {
+export default function NewCalendar(
+  { appointments }: { appointments: HealthWorkerAppointment[] },
+) {
   return (
     <div>
       <h2 className='text-base font-semibold leading-6 text-gray-900'>
-        Upcoming meetings
+        Upcoming appointments
       </h2>
       <div className='lg:grid lg:grid-cols-12 lg:gap-x-16'>
-        <div className='mt-10 text-center lg:col-start-8 lg:col-end-13 lg:row-start-1 lg:mt-9 xl:col-start-9'>
-          <div className='flex items-center text-gray-900'>
-            <button
-              type='button'
-              className='-m-1.5 flex flex-none items-center justify-center p-1.5 text-gray-400 hover:text-gray-500'
-            >
-              <span className='sr-only'>Previous month</span>
-              <ChevronLeftIcon className='h-5 w-5' aria-hidden='true' />
-            </button>
-            <div className='flex-auto text-sm font-semibold'>January</div>
-            <button
-              type='button'
-              className='-m-1.5 flex flex-none items-center justify-center p-1.5 text-gray-400 hover:text-gray-500'
-            >
-              <span className='sr-only'>Next month</span>
-              <ChevronRightIcon className='h-5 w-5' aria-hidden='true' />
-            </button>
-          </div>
-          <div className='mt-6 grid grid-cols-7 text-xs leading-6 text-gray-500'>
-            <div>M</div>
-            <div>T</div>
-            <div>W</div>
-            <div>T</div>
-            <div>F</div>
-            <div>S</div>
-            <div>S</div>
-          </div>
-          <div className='isolate mt-2 grid grid-cols-7 gap-px rounded-lg bg-gray-200 text-sm shadow ring-1 ring-gray-200'>
-            {days.map((day, dayIdx) => (
-              <button
-                key={day.date}
-                type='button'
-                className={classNames(
-                  'py-1.5 hover:bg-gray-100 focus:z-10',
-                  day.isCurrentMonth ? 'bg-white' : 'bg-gray-50',
-                  (day.isSelected || day.isToday) && 'font-semibold',
-                  day.isSelected && 'text-white',
-                  !day.isSelected && day.isCurrentMonth && !day.isToday &&
-                    'text-gray-900',
-                  !day.isSelected && !day.isCurrentMonth && !day.isToday &&
-                    'text-gray-400',
-                  day.isToday && !day.isSelected && 'text-indigo-600',
-                  dayIdx === 0 && 'rounded-tl-lg',
-                  dayIdx === 6 && 'rounded-tr-lg',
-                  dayIdx === days.length - 7 && 'rounded-bl-lg',
-                  dayIdx === days.length - 1 && 'rounded-br-lg',
-                )}
-              >
-                <time
-                  dateTime={day.date}
-                  className={classNames(
-                    'mx-auto flex h-7 w-7 items-center justify-center rounded-full',
-                    day.isSelected && day.isToday && 'bg-indigo-600',
-                    day.isSelected && !day.isToday && 'bg-gray-900',
-                  )}
-                >
-                  {day.date.split('-').pop().replace(/^0/, '')}
-                </time>
-              </button>
-            ))}
-          </div>
+        <Calendar>
           <button
             type='button'
             className='mt-8 w-full rounded-md bg-indigo-600 px-3 py-2 text-sm font-semibold text-white shadow hover:bg-indigo-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600'
           >
             Add event
           </button>
-        </div>
+        </Calendar>
         <ol className='mt-4 divide-y divide-gray-100 text-sm leading-6 lg:col-span-7 xl:col-span-8'>
-          {meetings.map((meeting) => (
-            <li
-              key={meeting.id}
-              className='relative flex space-x-6 py-6 xl:static'
-            >
-              <img
-                src={meeting.imageUrl}
-                alt=''
-                className='h-14 w-14 flex-none rounded-full'
-              />
-              <div className='flex-auto'>
-                <h3 className='pr-10 font-semibold text-gray-900 xl:pr-0'>
-                  {meeting.name}
-                </h3>
-                <dl className='mt-2 flex flex-col text-gray-500 xl:flex-row'>
-                  <div className='flex items-start space-x-3'>
-                    <dt className='mt-0.5'>
-                      <span className='sr-only'>Date</span>
-                      <CalendarIcon
-                        className='h-5 w-5 text-gray-400'
-                        aria-hidden='true'
-                      />
-                    </dt>
-                    <dd>
-                      <time dateTime={meeting.datetime}>
-                        {meeting.date} at {meeting.time}
-                      </time>
-                    </dd>
-                  </div>
-                  <div className='mt-2 flex items-start space-x-3 xl:ml-3.5 xl:mt-0 xl:border-l xl:border-gray-400 xl:border-opacity-50 xl:pl-3.5'>
-                    <dt className='mt-0.5'>
-                      <span className='sr-only'>Location</span>
-                      <MapPinIcon
-                        className='h-5 w-5 text-gray-400'
-                        aria-hidden='true'
-                      />
-                    </dt>
-                    <dd>{meeting.location}</dd>
-                  </div>
-                </dl>
-              </div>
-              <div className='absolute right-0 top-6 xl:relative xl:right-auto xl:top-auto xl:self-center'>
-                <div>
-                  <button
-                    type='button'
-                    className='-m-2 flex items-center rounded-full p-2 text-gray-500 hover:text-gray-600'
-                    id='menu-0-button'
-                    aria-expanded='false'
-                    aria-haspopup='true'
-                  >
-                    <span className='sr-only'>Open options</span>
-                    <svg
-                      className='h-5 w-5'
-                      viewBox='0 0 20 20'
-                      fill='currentColor'
-                      aria-hidden='true'
-                    >
-                      <path d='M3 10a1.5 1.5 0 113 0 1.5 1.5 0 01-3 0zM8.5 10a1.5 1.5 0 113 0 1.5 1.5 0 01-3 0zM15.5 8.5a1.5 1.5 0 100 3 1.5 1.5 0 000-3z' />
-                    </svg>
-                  </button>
-                </div>
-
-                {
-                  /* Dropdown menu, show/hide based on menu state.
-
-                  Entering: "transition ease-out duration-100"
-                    From: "transform opacity-0 scale-95"
-                    To: "transform opacity-100 scale-100"
-                  Leaving: "transition ease-in duration-75"
-                    From: "transform opacity-100 scale-100"
-                    To: "transform opacity-0 scale-95" */
-                }
-                <div
-                  className='absolute right-0 z-10 mt-2 w-36 origin-top-right rounded-md bg-white shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none'
-                  role='menu'
-                  aria-orientation='vertical'
-                  aria-labelledby='menu-0-button'
-                  tabindex='-1'
-                >
-                  <div className='py-1' role='none'>
-                    {/* <!-- Active: "bg-gray-100 text-gray-900", Not Active: "text-gray-700" --> */}
-                    <a
-                      href='#'
-                      className='text-gray-700 block px-4 py-2 text-sm'
-                      role='menuitem'
-                      tabIndex={-1}
-                      id='menu-0-item-0'
-                    >
-                      Edit
-                    </a>
-                    <a
-                      href='#'
-                      className='text-gray-700 block px-4 py-2 text-sm'
-                      role='menuitem'
-                      tabIndex={-1}
-                      id='menu-0-item-1'
-                    >
-                      Cancel
-                    </a>
-                  </div>
-                </div>
-              </div>
-            </li>
+          {appointments.map((appointment) => (
+            <Appointment appointment={appointment} />
           ))}
         </ol>
       </div>
