+++ external-clients/whatsapp.ts
@@ -14,6 +14,18 @@ const postMessageRoute = `https://graph.facebook.com/v17.0/${
 }/messages`
 const Authorization = `Bearer ${Deno.env.get('WHATSAPP_BEARER_TOKEN')}`
 
+export async function getFoo() {
+  const response = await fetch(
+    `https://graph.facebook.com/v17.0/12032535603/media`,
+    {
+      method: 'get',
+      headers: { Authorization, 'Content-Type': 'application/json' },
+    },
+  )
+
+  return response.json()
+}
+
 export function sendMessage({
   message,
   phone_number,
