+++ fresh.gen.ts
@@ -16,10 +16,11 @@ import * as $9 from './routes/index.tsx'
 import * as $10 from './routes/logged-in.tsx'
 import * as $11 from './routes/logout.tsx'
 import * as $12 from './routes/patient-reset.ts'
-import * as $$0 from './islands/date-picker.tsx'
-import * as $$1 from './islands/month-picker.tsx'
-import * as $$2 from './islands/set-availability-form.tsx'
-import * as $$3 from './islands/year-picker.tsx'
+import * as $$0 from './islands/appointment-menu.tsx'
+import * as $$1 from './islands/date-picker.tsx'
+import * as $$2 from './islands/month-picker.tsx'
+import * as $$3 from './islands/set-availability-form.tsx'
+import * as $$4 from './islands/year-picker.tsx'
 
 const manifest = {
   routes: {
@@ -38,10 +39,11 @@ const manifest = {
     './routes/patient-reset.ts': $12,
   },
   islands: {
-    './islands/date-picker.tsx': $$0,
-    './islands/month-picker.tsx': $$1,
-    './islands/set-availability-form.tsx': $$2,
-    './islands/year-picker.tsx': $$3,
+    './islands/appointment-menu.tsx': $$0,
+    './islands/date-picker.tsx': $$1,
+    './islands/month-picker.tsx': $$2,
+    './islands/set-availability-form.tsx': $$3,
+    './islands/year-picker.tsx': $$4,
   },
   baseUrl: import.meta.url,
   config,
